Modify user_model.go to split into signup_user_model.go and login_user_model.go
