+++ user-rest-service/Makefile
@@ -16,4 +16,4 @@ lint:
 
 .PHONY: test
 test:
-	GO_ENV=local LOCAL_SERVER_PORT=8080 LOCAL_CORS_ALLOWED_ORIGINS="http://localhost:3000" LOCAL_COOKIE_DOMAIN="localhost" LOCAL_COOKIE_SECURE=false LOCAL_ACCOUNT_HOST="localhost" LOCAL_ACCOUNT_PORT=8081 LOCAL_MYSQL_DSN="root:root@tcp(127.0.0.1:3306)/test_db" LOCAL_REDIS_DSN="127.0.0.1:6379" go test -race -v ./...
+	env GO_ENV=local LOCAL_SERVER_PORT=8080 LOCAL_CORS_ALLOWED_ORIGINS="http://localhost:3000" LOCAL_COOKIE_DOMAIN="localhost" LOCAL_COOKIE_SECURE=false LOCAL_ACCOUNT_HOST="localhost" LOCAL_ACCOUNT_PORT=8081 LOCAL_MYSQL_DSN="root:root@tcp(127.0.0.1:3306)/test_db" LOCAL_REDIS_DSN="127.0.0.1:6379" go test -race -v ./...
